bullet: Bounce on slopes!
There is a little guestimation that I have put here because I spent a good while
trying to fine tune the hell out of this. It should function like it would in
other games, but it'll still do a large bouce if you shoot from a far up
distance.

Before, this was actually just not handled at all.

+++ src/object/bullet.hpp
@@ -53,7 +53,7 @@ public:
 private:
   Player& m_player;
   Physic physic;
-  int life_count;
+  float life_count;
   SpritePtr sprite;
   SpritePtr lightsprite;
   BonusType type;
